move examples under src/test; drop the WAR stuff
