--- java/example/war/WEB-INF/appengine-web.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<appengine-web-app xmlns="http://appengine.google.com/ns/1.0">
-  <application>java-pipeline</application>
-  <version>1</version>
-
-  <runtime>java17</runtime> <!-- or another supported version -->
-  <app-engine-apis>true</app-engine-apis>
-
-  <!-- Configure java.util.logging -->
-  <system-properties>
-    <property name="com.google.appengine.tools.pipeline.BASE_URL" value="/pipeline/"/>
-    <property name="java.util.logging.config.file" value="WEB-INF/logging.properties"/>
-  </system-properties>
-
-</appengine-web-app>
