--- java/example/war/WEB-INF/functions.tld
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<taglib 
-    xmlns="http://java.sun.com/xml/ns/javaee"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-jsptaglibrary_2_1.xsd"
-    version="2.1">
-
-    <display-name>Custom Functions</display-name>    
-    <tlib-version>1.0</tlib-version>
-    <uri>http://github.com/GoogleCloudPlatform/appengine-pipelines/functions</uri>
-
-    <function>
-        <name>baseUrl</name>
-        <function-class>com.google.appengine.tools.pipeline.impl.servlets.PipelineServlet</function-class>
-        <function-signature>java.lang.String baseUrl()</function-signature>
-    </function>
-</taglib>
