--- java/example/war/index.html
@@ -1,34 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
-<!-- The HTML 4.01 Transitional DOCTYPE declaration-->
-<!-- above set at the top of the file will set     -->
-<!-- the browser's rendering engine into           -->
-<!-- "Quirks Mode". Replacing this declaration     -->
-<!-- with a "Standards Mode" doctype is supported, -->
-<!-- but may lead to some differences in layout.   -->
-
-<html>
-<head>
-    <meta http-equiv="content-type" content="text/html; charset=UTF-8">
-    <title>App Engine Pipeline Framework Demonstrations</title>
-</head>
-
-<body>
-<h1>Java App Engine Pipeline Framework</h1>
-
-<a href="javadoc/index.html">javadoc</a>
-
-<p></p>
-
-<table>
-    <tr>
-        <td colspan="2" style="font-weight:bold;">Demos</td>
-    </tr>
-    <tr>
-        <td><a href="gcd.jsp">Recursive GCD Calculation</a></td>
-    </tr>
-    <tr>
-        <td><a href="lettercount.jsp">Letter count. Spawns a new sub-job for each word.</a></td>
-    </tr>
-</table>
-</body>
-</html>
